requirements.txt
openai-agents
requests
python-dotenv
rich
